% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_data_dict.R
\docType{data}
\name{data_dict_col_names}
\alias{data_dict_col_names}
\title{Usual data dictionaries}
\format{
An object of class \code{list} of length 18.
}
\usage{
data_dict_col_names
}
\description{
Usual data dictionaries
}
\keyword{datasets}
